Split out common processors test and fixed nits in PR feedback.
